requirements.txt
asgiref==3.9.1
Django==4.2
djangorestframework==3.16.1
psycopg2-binary==2.9.10
python-dotenv==1.1.1
sqlparse==0.5.3
